Move delete button to form (WIP)
Initial values are wrong for edit form.
